Move common.py into its own package
